Revert "주여, 드디어 배포 갑니다."
