--- src/components/defaultmodal/portal.js
@@ -1,8 +1,0 @@
-import reactDom from 'react-dom';
-
-const ModalPortal = ({ children }) => {
-  const el = document.getElementById("modal");
-  return reactDom.createPortal(children, el);
-};
-
-export default ModalPortal;
