+++ src/components/dropdown/DropDown.jsx
@@ -1,4 +1,4 @@
-import { DivDropDownBox, DivMenuBox, DivCategoryBox, DivMenuLine } from './style';
+import { DivDropDownBox, DivMenuBox, DivCategoryBox } from './style';
 import { useState } from 'react'
 
 export default function DropDown({children, menus, size, alignRight=false}){
@@ -9,26 +9,19 @@ export default function DropDown({children, menus, size, alignRight=false}){
   const toggleOpen = () => {
     setIsOpen(!isOpen)
   }
-
-  const closeMenu = (menuFunction) => {
-    menuFunction()
-    setIsOpen(false)
-  }
-  
   return <>
   <a href="#" onClick={toggleOpen}>{children}</a>
 
   <DivDropDownBox isOpen={isOpen} size={size} alignRight={alignRight}>
     { menus.map((menuForder, index) => {
-      const menuKeys = new Array(menuForder.menu.length).fill(1).map((key, index) => key + index);
-      return <DivMenuLine style={{flexDirection: 'column'}} key={keys[index]}>{ menuForder.category ? <DivCategoryBox>{menuForder.category}</DivCategoryBox> : <div></div> }
+      const menuKeys = new Array(menuForder.menu.length).fill(1).map((key, index) => key + index)
+      return <>{ menuForder.category ? <DivCategoryBox key={keys[index]}>{menuForder.category}</DivCategoryBox> : <></>}
         {menuForder.menu.map((oneMenu, index) => {
-          return <a href={oneMenu.href || '#'} onClick={() => closeMenu(oneMenu.onClick)} key={menuKeys[index]}><DivMenuBox>{oneMenu.title}</DivMenuBox></a>
+          return <a href={oneMenu.href} key={menuKeys[index]}><DivMenuBox>{oneMenu.title}</DivMenuBox></a>
         })}
-        </DivMenuLine>
+        </>
       })
     }
   </DivDropDownBox>
   </>
 }
-
