--- src/components/loading/Loading.jsx
@@ -1,17 +1,0 @@
-import './spinning.css';
-import styled from 'styled-components';
-
-export default function Loading(){
-
-  return <DivSpinningContainer><div className="loadingio-spinner-spin-l7mkd54ox9j"><div className="ldio-5b1l7tpdv2q">
-    <div><div></div></div><div><div></div></div><div><div></div></div><div><div></div></div><div><div></div></div><div><div></div></div><div><div></div></div><div><div></div></div>
-  </div></div>
-  </DivSpinningContainer>
-
-}
-
-const DivSpinningContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`
