--- src/components/pin/PinStyle.jsx
@@ -1,126 +1,0 @@
-// import styled from 'styled-components';
-
-// export const PinBox = styled.div`
-//   width: 252px;
-//   height: 512px;
-//   border-radius: 16px;
-//   position: relative;
-//   overflow: hidden;
-//   background-color: green;
-// `;
-
-// export const PinTitle = styled.div`
-//   width: 0;
-//   height: 0;
-//   opacity: 0;
-// `;
-
-// export const PinModal = styled.div`
-//   width: 100%;
-//   height: 100%;
-//   position: absolute;
-//   top: 50%;
-//   left: 50%;
-//   transform: translate(-50%, -50%);
-
-//   opacity: 0;
-//   background-color: rgba(0, 0, 0, 0.1);
-//   transition-duration: 0.3s;
-//   &:hover {
-//     opacity: 1;
-//   }
-// `;
-
-// export const ModalHead = styled.div`
-//   width: 100%;
-//   height: 15%;
-//   position: fixed;
-//   top: 0;
-
-//   display: flex;
-//   justify-content: flex-end;
-//   align-items: center;
-// `;
-
-// export const ModalFoot = styled.div`
-//   width: 100%;
-//   height: 15%;
-//   position: fixed;
-//   bottom: 0;
-
-//   display: flex;
-//   justify-content: space-evenly;
-//   align-items: center;
-// `;
-
-// export const PinIconContainer = styled.div`
-//   width: 32px;
-//   height: 32px;
-//   border-radius: 50%;
-
-//   display: flex;
-//   justify-content: center;
-//   align-items: center;
-//   flex-shrink: 0;
-
-//   background-color: #f0f0f0;
-//   cursor: pointer;
-
-//   .pin-icon {
-//     width: 50%;
-//   }
-// `;
-
-// export const SavePin = styled.div`
-//   margin-right: 20px;
-//   width: 53px;
-//   height: 40px;
-//   border-radius: 22px;
-
-//   display: flex;
-//   justify-content: center;
-//   align-items: center;
-
-//   font-size: 14px;
-//   font-weight: 700;
-
-//   color: white;
-//   background-color: #e60023;
-//   cursor: pointer;
-// `;
-
-// export const Destination = styled.div`
-//   width: 132px;
-//   height: 32px;
-//   border-radius: 18px;
-
-//   display: flex;
-//   align-items: center;
-
-//   font-size: 16px;
-//   font-weight: 700;
-
-//   background-color: #f0f0f0;
-//   cursor: pointer;
-
-//   span {
-//     overflow: hidden;
-//   }
-// `;
-
-// export const PinImage = styled.div`
-//   width: 100%;
-//   height: 100%;
-//   /* position: absolute;
-//   top: 50%;
-//   left: 50%;
-//   transform: translate(-50%, 50%); */
-
-//   display: flex;
-//   justify-content: center;
-//   align-items: center;
-
-//   img {
-//     max-height: 100%;
-//   }
-// `;
