+++ src/index.js
@@ -1,20 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
-import axios from "axios";
-
-import store from "./redux/config/configStore";
-import { Provider } from "react-redux";
-
-axios.defaults.withCredentials = true;
-
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <Provider store={store}>
+  <React.StrictMode>
     <App />
-  </Provider>
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
