+++ src/pages/Write.jsx
@@ -1,20 +1,5 @@
-import React from 'react';
-import Header from '../components/header/header/Header';
-import PinWrite from '../components/pin/PinWrite';
-import { useCookies } from 'react-cookie';
-import LoginedHeader from '../components/header/loginedheader/LoginedHeader';
+export default function Write(){
 
-const Write = () => {
-  const [ cookie] = useCookies();
 
-  const isLogined= cookie['refresh_token'] && cookie['access_token'];
-
-  return (
-    <>
-      {isLogined ? <LoginedHeader /> :<Header />}
-      <PinWrite />
-    </>
-  );
-};
-
-export default Write;
+  return <h1>Write</h1>
+}
