--- src/pages/main.css
@@ -1,39 +1,0 @@
-#app { height: 100% }
-html,
-body {
-  position: relative;
-  height: 100%
-}
-
-.swiper {
-  width: 100vw;
-  height: 92vh;
-}
-
-.swiper-slide {
-  text-align: center;
-  font-size: 18px;
-  background: rgb(146, 114, 114);
-
-  /* 슬라이드 안 텍스트를 가운데 정렬 */
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: -webkit-flex;
-  display: flex;
-  -webkit-box-pack: center;
-  -ms-flex-pack: center;
-  -webkit-justify-content: center;
-  justify-content: center;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  -webkit-align-items: center;
-  align-items: center;
-}
-
-.swiper-slide img {
-  display: block;
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  position: relative;
-}
