+++ src/redux/config/configStore.js
@@ -1,14 +1,8 @@
-import { configureStore, getDefaultMiddleware } from '@reduxjs/toolkit';
-import PinSlice from '../module/PinSlice';
-import CommentSlice from '../module/CommentSlice';
+import { configureStore } from "@reduxjs/toolkit";
+
 
 const store = configureStore({
-  reducer: {
-    PinSlice, CommentSlice
-  },
-  middleware: getDefaultMiddleware({
-    serializableCheck: false,
-  }),
-});
+
+}, )
 
 export default store;
