--- src/shared/cookie.js
@@ -1,19 +1,0 @@
-import { Cookies } from "react-cookie";
-
-//생성자로 쿠키 생성
-const cookies = new Cookies();
-
-// setCookie : 로그인 시 사용자 정보를 담은 쿠키 생성
-export const setCookie = (name, value, option) => {
-  return cookies.set(name, value, { ...option });
-};
-
-// getCookie : 사용자 인증이 필요한 데이터를 요청할 때 쿠키 가져오기
-export const getCookie = (name) => {
-  return cookies.get(name);
-};
-
-// 로그아웃 시 쿠키 지우기
-export const removeCookie = (name) => {
-  return cookies.remove(name);
-};
